Extract editor command preview helpers
